Added more endpoints

+++ controller/customer.js
@@ -1,11 +1,21 @@
 const models = require('../models/index');
 const Customer = models.Customer;
+const Staff = models.Staff;
 const { validationResult } = require('express-validator');
 
 exports.getAll = async (req, res) => {
+	const staffId = req.staff._id;
+
+	//check if the staff making the request is still a registered staff
+	const registeredStaff = await Staff.findById(staffId);
+	if (!registeredStaff)
+		return res
+			.status(401)
+			.json({ message: 'Sorry, you do not have access to this route' });
+
 	try {
 		const customers = await Customer.find()
-			.populate('washes', 'washDate washId')
+			.populate('washes', 'washDate washId number_of_wash')
 			.populate('payments', 'payment_date amount payment_mode');
 		if (customers.length <= 0) {
 			return res.status(404).json({
@@ -26,6 +36,16 @@ exports.getAll = async (req, res) => {
 
 exports.register = async (req, res) => {
 	const { customer_name, email, mobile_num, address } = req.body;
+
+	const staffId = req.staff._id;
+
+	//check if the staff making the request is still a registered staff
+	const registeredStaff = await Staff.findById(staffId);
+	if (!registeredStaff)
+		return res
+			.status(401)
+			.json({ message: 'Sorry, you do not have access to this route' });
+
 	//get the validation results
 	const errors = validationResult(req);
 
@@ -43,11 +63,14 @@ exports.register = async (req, res) => {
 				'Customer with the same email as been registered... try another email or update existing customer',
 		});
 
+	//modify the mobile number to a nigerian number
+	modifiedNum = `(+234)-${mobile_num}`;
+
 	const name = customer_name.toLowerCase();
 	const user = new Customer({
 		customer_name: name,
 		email,
-		mobile_num,
+		mobile_num: modifiedNum,
 		address,
 	});
 
@@ -67,6 +90,15 @@ exports.register = async (req, res) => {
 
 exports.getOne = async (req, res) => {
 	const id = req.params.id;
+	const staffId = req.staff._id;
+
+	//check if the staff making the request is still a registered staff
+	const registeredStaff = await Staff.findById(staffId);
+	if (!registeredStaff)
+		return res
+			.status(401)
+			.json({ message: 'Sorry, you do not have access to this route' });
+
 	try {
 		const customer = await Customer.findById(id)
 			.populate('washes', 'washDate washId')
@@ -91,6 +123,15 @@ exports.getOne = async (req, res) => {
 
 exports.updateOne = async (req, res) => {
 	const id = req.params.id;
+	const staffId = req.staff._id;
+
+	//check if the staff making the request is still a registered staff
+	const registeredStaff = await Staff.findById(staffId);
+	if (!registeredStaff)
+		return res
+			.status(401)
+			.json({ message: 'Sorry, you do not have access to this route' });
+
 	try {
 		const customer = await Customer.findByIdAndUpdate(id, req.body, {
 			new: true,
@@ -113,6 +154,14 @@ exports.updateOne = async (req, res) => {
 
 exports.deleteOne = async (req, res) => {
 	const id = req.params.id;
+	const staffId = req.staff._id;
+
+	//check if the staff making the request is still a registered staff
+	const registeredStaff = await Staff.findById(staffId);
+	if (!registeredStaff)
+		return res
+			.status(401)
+			.json({ message: 'Sorry, you do not have access to this route' });
 	try {
 		const deletedCustomer = await Customer.findByIdAndRemove(id, {
 			useFindAndModify: false,
@@ -132,6 +181,14 @@ exports.deleteOne = async (req, res) => {
 };
 
 exports.deleteAll = async (req, res) => {
+	const staffId = req.staff._id;
+
+	//check if the staff making the request is still a registered staff
+	const registeredStaff = await Staff.findById(staffId);
+	if (!registeredStaff)
+		return res
+			.status(401)
+			.json({ message: 'Sorry, you do not have access to this route' });
 	try {
 		const deletedCustomers = await Customer.deleteMany({});
