+++ controller/wash.js
@@ -2,9 +2,19 @@ const models = require('../models/index');
 const Staff = models.Staff;
 const Wash = models.Wash;
 const Customer = models.Customer;
+const { validationResult } = require('express-validator');
 
 //get all washes
 exports.allWashes = async (req, res) => {
+	const staffId = req.staff._id;
+
+	//check if the staff making the request is still a registered staff
+	const registeredStaff = await Staff.findById(staffId);
+	if (!registeredStaff)
+		return res
+			.status(401)
+			.json({ message: 'Sorry, you do not have access to this route' });
+
 	try {
 		const washes = await Wash.find()
 			.populate('payment', 'payment_date payment_mode amount')
@@ -28,10 +38,93 @@ exports.allWashes = async (req, res) => {
 	}
 };
 
+//get wash by Id
+exports.getWash = async (req, res) => {
+	const id = req.params.id;
+	const staffId = req.staff._id;
+
+	//check if the staff making the request is still a registered staff
+	const registeredStaff = await Staff.findById(staffId);
+	if (!registeredStaff)
+		return res
+			.status(401)
+			.json({ message: 'Sorry, you do not have access to this route' });
+
+	try {
+		const wash = await Wash.findOne({ _id: id });
+		if (!wash) {
+			return res.status(422).json({
+				message: 'There is no wash record with such id',
+			});
+		} else {
+			return res.status(200).json({
+				message: 'Found wash record',
+				wash: wash,
+			});
+		}
+	} catch (error) {
+		return res.status(500).json({
+			message: 'Error in getting wash record',
+			error: error.message,
+		});
+	}
+};
+
+//search wash record by a DLSwashID
+exports.getWashByDlsId = async (req, res) => {
+	const washId = req.body.washId;
+	const staffId = req.staff._id;
+
+	//check if the staff making the request is still a registered staff
+	const registeredStaff = await Staff.findById(staffId);
+	if (!registeredStaff)
+		return res
+			.status(401)
+			.json({ message: 'Sorry, you do not have access to this route' });
+
+	const errors = validationResult(req);
+
+	if (!errors.isEmpty())
+		return res.status(422).json({
+			errors: errors.array(),
+		});
+
+	try {
+		const wash = await Wash.findOne({ washId: washId })
+			.populate('payment', 'payment_date payment_mode amount')
+			.populate('customer', 'email address')
+			.populate('staff', 'name');
+
+		if (!wash)
+			return res.status(404).json({
+				message: 'Invalid Wash id',
+			});
+
+		return res.status(200).json({
+			message: 'Found the wash record',
+			wash: wash,
+		});
+	} catch (error) {
+		return res.status(500).json({
+			message: 'Error in getting the wash',
+			error: error.message,
+		});
+	}
+};
+
 //new wash
 exports.newWash = async (req, res) => {
-	const { customer_name, number_of_wash } = req.body;
+	const { customer_name, number_of_wash, amount } = req.body;
 	const id = req.staff._id;
+
+	//get the validation results
+	const errors = validationResult(req);
+
+	if (!errors.isEmpty()) {
+		return res.status(422).json({
+			errors: errors.array(),
+		});
+	}
 
 	const name = customer_name.toLowerCase();
 
@@ -52,11 +145,13 @@ exports.newWash = async (req, res) => {
 
 	const date = new Date();
 	const washDate = date.toString();
+	const modifiedAmount = `\u20A6${amount.toString()}`;
 
 	const wash = new Wash({
 		customer_name: name,
 		washDate,
 		number_of_wash,
+		amount: modifiedAmount,
 	});
 
 	try {
@@ -82,7 +177,7 @@ exports.newWash = async (req, res) => {
 			);
 
 			return res.status(200).json({
-				message: 'Yayy!!! New wash has been made',
+				message: 'Yayy!!! New wash has been made. Copy the washId',
 				wash: updatedWash,
 			});
 		}
@@ -93,8 +188,50 @@ exports.newWash = async (req, res) => {
 	}
 };
 
+//delete wash by Id
+exports.deleteWash = async (req, res) => {
+	const id = req.params.id;
+	const staffId = req.staff._id;
+
+	//check if the staff making the request is still a registered staff
+	const registeredStaff = await Staff.findById(staffId);
+	if (!registeredStaff)
+		return res
+			.status(401)
+			.json({ message: 'Sorry, you do not have access to this route' });
+
+	try {
+		const deletedWash = await Wash.findByIdAndDelete(id, {
+			useFindAndModify: false,
+		});
+		if (!deletedWash) {
+			return res.status(422).json({
+				message: 'There is no wash record with such id',
+			});
+		} else {
+			return res.status(200).json({
+				message: 'Wash record has been deleted successfully',
+			});
+		}
+	} catch (error) {
+		return res.status(500).json({
+			message: 'Error in deleting wash record',
+			error: error.message,
+		});
+	}
+};
+
 //delete all washes
 exports.removeAllWashes = async (req, res) => {
+	const staffId = req.staff._id;
+
+	//check if the staff making the request is still a registered staff
+	const registeredStaff = await Staff.findById(staffId);
+	if (!registeredStaff)
+		return res
+			.status(401)
+			.json({ message: 'Sorry, you do not have access to this route' });
+
 	try {
 		const deletedWashes = await Wash.deleteMany({});
 		if (deletedWashes) {
